# controllers/workout.go
package controllers

import (
	"github.com/gin-gonic/gin"
)

// CreateNewNote godoc
// @Summary      Get All Equipment
// @Description  get all equipment
// @Tags         workout
// @Accept       json
// @Produce      json
// @Param        req  body      requestModel.PaginatedRequest true "Get All Equipment Request"
// @Success      200  {object}  responseModel.PaginatedResponse
// @Failure      400  {object}  responseModel.PaginatedResponse
// @Router       /workout [get]
// @Security     ApiKeyAuth
func GetAllWorkout(c *gin.Context) {

}

// CreateNewNote godoc
// @Summary      Get All Equipment
// @Description  get all equipment
// @Tags         workout
// @Accept       json
// @Produce      json
// @Param        req  body      requestModel.PaginatedRequest true "Get All Equipment Request"
// @Success      200  {object}  responseModel.PaginatedResponse
// @Failure      400  {object}  responseModel.PaginatedResponse
// @Router       /workout [post]
// @Security     ApiKeyAuth
func CreateWorkout(c *gin.Context) {

}

// CreateNewNote godoc
// @Summary      Get All Equipment
// @Description  get all equipment
// @Tags         workout
// @Accept       json
// @Produce      json
// @Param        req  body      requestModel.PaginatedRequest true "Get All Equipment Request"
// @Success      200  {object}  responseModel.PaginatedResponse
// @Failure      400  {object}  responseModel.PaginatedResponse
// @Router       /workout [put]
// @Security     ApiKeyAuth
func UpdateWorkout(c *gin.Context) {

}

// CreateNewNote godoc
// @Summary      Get All Equipment
// @Description  get all equipment
// @Tags         workout
// @Accept       json
// @Produce      json
// @Param        req  body      requestModel.PaginatedRequest true "Get All Equipment Request"
// @Success      200  {object}  responseModel.PaginatedResponse
// @Failure      400  {object}  responseModel.PaginatedResponse
// @Router       /workout [delete]
// @Security     ApiKeyAuth
func DeleteWorkout(c *gin.Context) {

}